backend/requirements.txt
fastapi[all]
uvicorn
pydantic
httpx
python-dotenv
requests
bey
